+++ src/main/java/logic/commands/delete/DeleteTaskCommand.java
@@ -1,11 +1,12 @@
-package logic.commands;
+package logic.commands.delete;
 
 import data.FileAccess;
-import gui.popups.WarningPopup;
+import gui.popups.notification.WarningPopup;
+import logic.commands.Command;
 import logic.treeItems.ProjectTreeItem;
 import logic.treeItems.TaskTreeItem;
 
-public class DeleteTaskCommand implements Command{
+public class DeleteTaskCommand implements Command {
 
     private final TaskTreeItem task;
